(#5) Arrumando conflitos
