+++ frontend/package.json
@@ -11,6 +11,7 @@
     "@types/react-dom": "^17.0.9",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.2.0",
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.1",
     "typescript": "^4.3.5",
